Fixed the way adept-utils and callpath were being compiled by downloading boost and pointing adept-utils to it.
The tools directory now has its own Makefile.

Generalized the measurement API to support multiple types of measurements. Time and PAPI counters are currently supported. Adjusted modules to work with the new API. The user of a module must now specify the measurements they wish to make. Added a dependency on PAPI as a result.