Fixing bugs in AMG
Fixed bug in merging where moduleTraceStreams that were not perfectly aligned were producing invalid traceIDs.

Modified hier_merge to produce Sight-formatted logs, along with infrastructure changes to make this possible.

Extended interface of compModule constructors to make them easier to use.

Added a new comparator that checks if two values are identical or not.

Added a grep application.

Added mpiclang and mpiclang++ (stolen from LC scripts) to make it easier to add the MPI libraries to the clang command line.